[Linux] Fix all errors and most warnings

+++ src/impl_core.cpp
@@ -721,10 +721,10 @@ namespace daxa
     auto to_string(ImageMipArraySlice slice) -> std::string
     {
         return fmt::format("mips: {}-{}, layers: {}-{}",
-                           static_cast<u32>(slice.base_mip_level),
-                           static_cast<u32>(slice.base_mip_level + slice.level_count - 1),
-                           static_cast<u32>(slice.base_array_layer),
-                           static_cast<u32>(slice.base_array_layer + slice.layer_count - 1));
+                           slice.base_mip_level,
+                           slice.base_mip_level + slice.level_count - 1,
+                           slice.base_array_layer,
+                           slice.base_array_layer + slice.layer_count - 1);
     }
 
     auto to_string(ImageArraySlice slice) -> std::string
